Joey Commit new project // updated code

+++ ICE-Server/ICE-Server/Repository/BuildingsRepository.cs
@@ -6,22 +6,26 @@ using System.Collections.Concurrent;
 using ICE_Server.Interfaces;
 using ICE_Server.Models;
 using ICE_Server.DAL;
-
+using System.Data.Entity.Infrastructure;
 
 namespace ICE_Server.Repository
 {
     public class BuildingsRepository : IRepository<Building>
     {
-        ICEContext context = new ICEContext();
+        private ICEContext context = new ICEContext();
+        public BuildingsRepository(ICEContext context)
+        {
+            this.context = context;
+        }
 
         public IEnumerable<Building> GetAll()
         {
             return context.Buildings;
         }
 
-        public Building Get(int ids)
+        public Building Get(int id)
         {
-            var result = (from r in context.Buildings where r.ID == ids select r).FirstOrDefault();
+            var result = (from r in context.Buildings where r.ID == id select r).FirstOrDefault();
             return result;
         }
 
@@ -34,15 +38,41 @@ namespace ICE_Server.Repository
 
         public bool Update(Building item, int[] ids)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Entry(item).State = System.Data.Entity.EntityState.Modified;
-            context.SaveChanges();
-            return true;
+
+            try
+            {
+                context.SaveChanges();
+                return true;
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                throw;
+
+            }
+
         }
         public bool Delete(Building item)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Buildings.Remove(item);
             context.SaveChanges();
+
             return true;
+        }
+
+        private bool checkEntry(int id)
+        {
+            return context.Buildings.Count(e => e.ID == id) > 0;
         }
