+++ ICE-Server/ICE-Server/Repository/DevicesRepository.cs
@@ -6,14 +6,18 @@ using System.Collections.Concurrent;
 using ICE_Server.Interfaces;
 using ICE_Server.Models;
 using ICE_Server.DAL;
-
+using System.Data.Entity.Infrastructure;
 
 namespace ICE_Server.Repository
 {
     public class DevicesRepository : IRepository<Device>
     {
-        ICEContext context = new ICEContext();
-        
+        private ICEContext context = new ICEContext();
+        public DevicesRepository(ICEContext context)
+        {
+            this.context = context;
+        }
+
         public IEnumerable<Device> GetAll()
         {
             return context.Devices;
@@ -34,15 +38,41 @@ namespace ICE_Server.Repository
 
         public bool Update(Device item, int[] ids)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Entry(item).State = System.Data.Entity.EntityState.Modified;
-            context.SaveChanges();
-            return true;
+
+            try
+            {
+                context.SaveChanges();
+                return true;
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                throw;
+
+            }
+
         }
         public bool Delete(Device item)
         {
+            if (checkEntry(item.ID) == false)
+            {
+                return false;
+            }
+
             context.Devices.Remove(item);
             context.SaveChanges();
+
             return true;
+        }
+
+        private bool checkEntry(int id)
+        {
+            return context.Devices.Count(e => e.ID == id) > 0;
         }
